refactor: migrate graphql-rewrite crate from cpython to pyo3
