+++ edb/graphql-rewrite/_graphql_rewrite.pyi
@@ -1,0 +1,11 @@
+from typing import *
+
+class Entry:
+    key: str
+    key_vars: List[str]
+    variables: Dict[str, Any]
+    substitutions: Dict[str, Tuple[str, int, int]]
+
+    def tokens(self) -> List[Tuple[Any, int, int, int, int, Any]]: ...
+
+def rewrite(operation: Optional[str], text: str) -> Entry: ...
